+++ acf-extensions.php
@@ -50,6 +50,7 @@ function include_acf_extensions_plugin() {
 	include_once('fields/class-acf-field-latest_posts.php');
 	include_once('fields/class-acf-field-better_link.php');
 	include_once('fields/class-acf-field-better_map.php');
+	include_once('fields/class-acf-field-better_image.php');
 
 	include_once('rules/class-acf-rule-multisite.php');
 	include_once('rules/class-acf-rule-tax-type.php');